Replace one() with indexes
Using one() causes this module to break when multiple CNAMEs are
provided, because if there are more than one DNS entries, this line
calls the one function with a list of length > 1. This causes an error,
as intended by the design of one():
https://developer.hashicorp.com/terraform/language/functions/one.

As best I can tell from the logic, we can be confident that those arrays
are non-empty (so there will always be an item at index 0) if those
statements are executed.